requirements.txt
discord.py-self
openai>=1.0.0
notion-client>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0

google-generativeai>=0.3.0
pillow>=10.0.0
